refactor: add usePersistedColumns
