+++ src/components/SystemsView/hooks/useColumns.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 import type { OnSort, SortDirection } from '../SystemsView';
 import {
   getColumnMinWidthStyle,
@@ -8,6 +8,7 @@ import {
 import { STICKY_ACTIONS_HEADER_PROPS } from '../utils/stickyActionsColumn';
 import { getStickyNameHeaderProps } from '../utils/stickyNameColumn';
 import defaultColumns, { type Column } from '../columns/allColumnDefinitions';
+import { usePersistedColumns } from './usePersistedColumns';
 
 export const INITIAL_SORT: {
   sortBy: Column['sortBy'];
@@ -25,103 +26,6 @@ const FALLBACK_SORT: {
   direction: 'asc',
 };
 
-const STORAGE_KEY = 'ui.systems-view.columns';
-
-type ColumnPref = {
-  key: string;
-  isShown: boolean;
-};
-
-const getColumnSnapshot = (columns: readonly Column[]): ColumnPref[] =>
-  columns.map((column) => ({
-    key: column.key,
-    isShown: column.isShown ?? column.isShownByDefault,
-  }));
-
-const loadColumnPrefs = (): ColumnPref[] | null => {
-  try {
-    const raw = localStorage.getItem(STORAGE_KEY);
-    if (!raw) {
-      return null;
-    }
-
-    const parsed: unknown = JSON.parse(raw);
-    if (!Array.isArray(parsed)) {
-      return null;
-    }
-
-    const prefs = parsed.filter(
-      (item): item is ColumnPref =>
-        typeof item === 'object' &&
-        item !== null &&
-        typeof (item as ColumnPref).key === 'string' &&
-        typeof (item as ColumnPref).isShown === 'boolean',
-    );
-
-    return prefs.length > 0 ? prefs : null;
-  } catch {
-    return null;
-  }
-};
-
-const saveColumnPrefs = (columns: readonly Column[]): void => {
-  try {
-    localStorage.setItem(
-      STORAGE_KEY,
-      JSON.stringify(getColumnSnapshot(columns)),
-    );
-  } catch {
-    // Ignore quota errors or storage unavailable (e.g. private browsing).
-  }
-};
-
-const mergeColumnPrefs = (
-  defaults: readonly Column[],
-  stored: ColumnPref[] | null,
-): readonly Column[] => {
-  if (!stored) {
-    return defaults;
-  }
-
-  const storedByKey = new Map(
-    stored.map((pref, index) => [
-      pref.key,
-      { isShown: pref.isShown, order: index },
-    ]),
-  );
-
-  const merged = defaults.map((column) => {
-    const pref = storedByKey.get(column.key);
-    if (pref) {
-      return { ...column, isShown: pref.isShown };
-    }
-
-    return column;
-  });
-
-  return [...merged].sort((a, b) => {
-    const aStored = storedByKey.get(a.key);
-    const bStored = storedByKey.get(b.key);
-
-    if (aStored && bStored) {
-      return aStored.order - bStored.order;
-    }
-
-    if (aStored) {
-      return -1;
-    }
-
-    if (bStored) {
-      return 1;
-    }
-
-    return (
-      defaults.findIndex((column) => column.key === a.key) -
-      defaults.findIndex((column) => column.key === b.key)
-    );
-  });
-};
-
 interface UseColumnParams {
   sortBy: Column['sortBy'];
   onSort: OnSort;
@@ -135,13 +39,7 @@ export const useColumns = ({
   direction,
   isInventoryViewsEnabled,
 }: UseColumnParams) => {
-  const [columns, setColumns] = useState<readonly Column[]>(() =>
-    mergeColumnPrefs(defaultColumns, loadColumnPrefs()),
-  );
-
-  useEffect(() => {
-    saveColumnPrefs(columns);
-  }, [columns]);
+  const { columns, setColumns } = usePersistedColumns(defaultColumns);
 
   const fromSortByToIndex = useCallback(
     (sortBy?: Column['sortBy']) =>
